divided eth and led functionality into seperate files
