+++ .gitignore
@@ -1,0 +1,4 @@
+.vscode
+.vscode/*
+build/
+build/*
